LangChain Helper
In this file we have created and used the different chains, one chain is created to generate the "restaurant name" and another chain is created to generate the "menu items", these both chains are connected through sequential chain, here the output of one call is used as the input to the next call. 
A sequential chain is a powerful feature in Langchain that allows user to combine multiple individual chains into a single, streamlined process. These individual chains, often called sub chains, perform specific tasks.